Fix issue with stale webdriver processes
Interrupting the test suite with Ctrl+C in the shell sends a signal to
all processes in the foreground group (this is the standard unix shell
behavior). By default `popen` starts processes in the same group, so
this means they will receive the SIGINT from the shell. However the main
process traps the signal ands sends a second SIGINT which results in an
abrupt termination of rspec (as opposed to a graceful shutdown when only
a single interrupt is received).

In particular if this happens while running feature specs the webdriver
processes can be left intact keeping network ports open which can cause
test failures in subsequent runs. (with very cryptic error messages)