Use half of processors by default
In a typical test run apart from the ruby processes we also have the
database, redis and chromium (for feature specs) doing non-trivial
amounts of work. This combined with the architectural complexity of
modern CPUs (e.g. hyper-threading) means we actually get worse results
due to contention when maxing out the ruby processes.

Note that getting the best results is still going to require some
experimentation on a particular machine, but this just seems like a more
sensible default for a typical developer machine.